Add threshold to TagSimilarityEncoder

+++ pipeline/io.py
@@ -167,9 +167,9 @@ class BBBinaryRepoSink(Sink):
                 db.xRotation = float(detection.orientations[j, 2])
                 db.localizerSaliency = float(detection.saliencies[j, 0])
                 db.radius = float(detection.radii[j])
-                descriptor = db.init('descriptor', len(detection.descriptors[j]))          
+                descriptor = db.init('descriptor', len(detection.descriptors[j]))
                 for k, part in enumerate(detection.descriptors[j]):
-                    descriptor[k] = float(part)
+                    descriptor[k] = int(part)
                 decodedId = db.init('decodedId', len(detection.ids[j]))
                 for k, bit in enumerate(detection.ids[j]):
                     decodedId[k] = int(round(255*bit))
